requirements.txt
streamlit
mitreattack-python
matplotlib